version 1.2 - misc bug fixes
NOTE: database is incompatible, since the series names are
      canonicalized differently

load is now not noisy for --cron

We try harder to make the filename and torrent description match, so
that means canonicalizing the series name from the description, since
many of the characters that might appear there can't appear in the
filename.  Also, make filename parsing case insensitive.  Also, the
removal of "(uncut)" from the filename was coded wrong, so it
prevented most matching.

Have a new slot for the "pretty" series name, which is used when
humans will see it.

Add install_config makefile rule that copies config.cl to $(DESTDIR).

"make all" does testing now.

Config file:
- capitalize all series names
- new series "James May's Man Lab"