version 1.17 - now with fuzzy logic!
Fuzzy comparison of series names extracted from the `description'
element in the RSS feed and the filename extracted from the RSS feed.
Basically, I now construct the perfect series name while comparing
both of them.

Refactored the feed groking code when bring back the eztv.it feed
parser.  The parsers have been unified a little and are simpler.

rssreader: export *uri-to-package* so tget can augment it.

Add tests for fuzzy matching.  At this point, the test feeds only
throw a single warning on series names (and the data is bad).