version 1.20 replace `archive/restore' with `compact'
The acache save-database/restore-database isn't meant to be used
across databases with schema changes.  That is, you can't
save-database and then later restore-database into a version of the
program that has a different schema.  For this reason, do away with
the general --archive and --restore arguments and just have a
--compact argument, which does the right thing.  This is actually much
easier to implement.  --backup-method is also gone.  And --auto-backup
"restore" is gone and "compact" is here.

Schema bumped to 7.

Another major change: for series like The Daily Show that use dates
for episode numbers, have a pretty-epnum slot in the episode class
that is used for printing.  S2013E86 isn't as recognizable as
2013.03.28.  Now, the latter is printed when presenting to humans.
This was actually the change I was trying to do.  The archive/restore
came up in testing.