version 1.26 - download .torrent files to a directory
deftransmission has been replaced with set-torrent-handler, which
takes either an object returned by make-transmission-remote or a
pathname.  In the latter case, it will download the .torrent file to a
directory rather than have transmission-remote handle it.  This latter
capability is untested.

The on deftransmission-style of handling the torrents still works, but
it will be removed at some point.