version 1.30 - fix rss logging
First, reset the rss log each run.  That way it makes it easier to
decide to save it.  Second, don't save the cached rss feed, since we
might have to visit it multiple times, based on the config file.  This
way, the file is even smaller.

Update public config file, too.