v2.4: better handle date-based episodes
The "complete-to" mechanism is great for seasonal episodic shows, but
not good for date-based shows, like The Daily Show and The Colbert
Report.  When episodes come out of order, shows can be missed.

There's a new defseries option, :date-based, which if `t', turns off
the complete-to mechanism for this series.

Database schema version 10, to add a slot to the series object.