New (lisp) version of seedstatus program
This one uses Ahmon's bittorrent parser so I can determine the
tracker associate with a torrent, so I can implement special seeding
rules for torrents based on the tracker.

The seedstatus program is installed in /usr/local/lib/seedstuff with a
symlink in /usr/local/bin/seedstatus.

Also removed the rpm build of tget from the Makefile.  It was never
used and added a lot of cruft.