v2.8.0: add manual mode
It is convenient, like when TVT is down, to be able to download torrents
manually, but still keep track of which episodes you've downloaded.
Normally, you would need to use the --catch-up-series to manually do
that.  The new --add argument manually adds .torrent files from the
given directory and sends them to your torrent client by whatever
means you have defined in your config file.

This new mode uses the `bittorrent' repo
  https://github.com/dancysoft/bittorrent
which is used to extract information inside the torrent file, such as
pub-date and length.  Originally, I used the filename inside the
.torrent, but that proved less reliable because of typos.  Now, I use
the .torrent filename itself, which can be adjusted if there is a
need.  In the future, bittorrent might not be a requirement.