v2.9.1: tests pass once again
After fixing the feed parser for BTN, it exposed a problem in my
config file, which made the test suite fail.  The config file has been
adjusted.

Another change to the config file: my-quality looks at
episode-torrent-url and accepts different quality for BTN.

The TVT feed parser aborts when it sees a season, complete month or
special.  tget isn't meant to handle those anyway.  It also makes
parsing the ambiguous "Archer.2009.S04...." not something I have to
deal with.  :)

The above alluded to above in the BTN parser: the :case-fold argument
to DOES NOT WORK.  Workaround that bug.