seedstatus: handle renamed torrents
Sometimes the name is so bad from the site (looking at you BTN!)
that I need to rename it manually for Plex to see it.  In these cases
we couldn't tell what the tracker was, but there's a hint in the
.torrent files that Transmission saves: it includes a subset of the
hash at the end of the filename.  We can use this to match differing
filenames.