tcleanup: comments + fix + feature
The fix is to downcase the container before converting it to a
keyword.  Sometimes they're in upper case.

The comments are TODOs and additions.

Also, removing warning when a torrent is finished and the
torrent-seeded-for is null.  Public trackers don't fill in this info.

The feature: add --remove and -h command line args.