tcleanup: fix transmission-filename-to-tracker
Test that the torrent file exists before operating on it.  For some
reason I was seeing it disappear between the `directory' and
bdecode-file opening it.  Installed a warning, which I've not seen for
a while.  The machine transmission runs on was behaving really
strangely and I rebooted it, so maybe it was related to that.