New program: plexfix
plexfix adds symlinks to files that Plex Media Server cannot see
because their names are malformed.  Because we might be seeding the
file with the "bad" name, we can't just rename it.  So, add a symlink
to it so PMS will see it.

This program is called when a torrent finishes downloaded via
Transmission.  It gets the torrent filename from the environment, in
the variables

  TR_TORRENT_DIR
  TR_TORRENT_NAME