tcleanup: handle multiple files in a torrent
For the cleanup files phase (tcleanup-files), seedingp wasn't correct
for torrents with multiple files.  In these cases, the torrent "Name"
points to a directory.  We need to process all the entries in that
directory so we know we are seeding them.  Otherwise, we could remove
files we watch but are still seeding.