v4.4.0: new deftracker keyword: upload-limit
Docs for upload-limit:

  set the upload limit for this tracker.  This is useful for when the
  tracker unreliably gives you upload credit for what you upload (I'm
  looking at you TvT!).  Sometimes, if you upload too fast the tracker
  will not count a significant amount of your upload data.

Make test suite less hardwired to data.  Also, test suite was out of
sync with reality, for some reason.  It's like I fixed a bug in the
core program that made the test suite out of date.  Even though I
don't know why that was, I fixed it (the changes look correct).