v5.5.0: fix two bugs
The first one is a serious and old bug.  Don't know how this didn't
come up before.

Bug 1:

The cleanup phase wasn't recognizing a season pack from BTN.  It was
because the tracker info wasn't available at the time the cleanup info
was printed.

New function defs.cl:tracker-name-to-instance with turns a string name
for the tracker (different than the keyword "name" defined with the
tracker!!) into the tracker instance.

In the tracker object, I renamed "tracker" to "tracker-name" and added
a new "tracker-instance".  That's much clearer.

When initializing the list of torrents, set torrent-tracker-instance.

Bug 2:

The RE being used in transmission.cl:transmission-filename-to-tracker
was flat wrong.  The names of the files are "${hash}.torrent" and the
RE was expecting something before the ${hash}.