tget.cl: add a WIP hack to deal with manual adds
If a series has a ":" in the name but the filename does not, then
adding it will not find the series.  Quick fix is to add an alias for
it.  The real fix would be to fix fuzzy-compare-series-names to return
nil when there isn't really a match, but that code is impossible to
understand, even though it has a few comments.